Created nav menu block.
