+++ frontend/containers/App/App.css
@@ -11,3 +11,4 @@ body {
 .RightColumn {
     flex: 0 0 auto;
 }
+
