+++ frontend/containers/Frig/Frig.css
@@ -1,0 +1,1 @@
+.FrigContainer {}
