+++ frontend/containers/Menu/Menu.css
@@ -1,0 +1,1 @@
+.MenuContainer {}
